backend/requirements/base.txt
django>=5.0.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-environ>=0.11.2
psycopg2-binary>=2.9.9
redis>=5.0.1
channels[daphne]>=4.0.0
celery>=5.3.6
django-celery-results>=2.5.1
anthropic>=0.20.0
pydantic>=2.6.4
Pillow>=10.2.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
uvicorn>=0.27.1
whitenoise>=6.6.0
django-filter>=24.2
drf-spectacular>=0.27.1
